+++ src/main/java/za/co/photo_sharing/app_ws/shared/dto/SkillSetDto.java
@@ -12,5 +12,5 @@ public class SkillSetDto {
     private long id;
     private String skill;
     private double rating;
-    private double ratingCalc;
+    private int ratingCalc;
 }
